Invert dependency on spec namespace, allow initialization instead
This means we don't strictly need spec or expound as a dependency unless we
explicitly enable it. This can be quite a saving in compiled assets (100kb
pprint.cljs, 60kb expound, ~50kb spec).

Also means we may optionally be able to introduce non-spec equivalents (e.g.
malli).

There are some pieces of code in this commit which could be refactored or
improved but the intent is currently just to realign the namespace dependency
and introduce the ability to optionally initialize spec validation.